updating documentation, moving documentation snippet into its own test file
